Improve performance of sparse copy
First, we remove the `lseek` that happens in each iteration of the loop
which causes a flush. Instead of finding the next hole in each iteration,
we compute it when we find the data section. This gives a 10x improvement
in performance.

Second, we use `f_iosize` (or user specified block size) for the read/write
buffer. This improves performance by another 4x. Careful consideration is
made to ensure the following cases are still correct:

1. If a single read buffer completely encloses a hole, then we waste some
   cycles writing 00s but the hole will still be punched in the end.
2. If a single read buffer partially overlaps with a hole, then same as
   above, but the remaining hole will be skipped with lseek.

A minor bug is also fixed from the original code which never checks the
return value of `nread` so if the read fails, incorrect data could be
copied.